refactor(store): move selector utils to selector folder
