+++ packages/store/src/selectors/create-property-selectors.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '../utils/selector-utils';
+import { createSelector } from './selector-utils';
 import { ensureValidSelector } from './selector-checks.util';
 import { SelectorDef } from './selector-types.util';
