+++ packages/store/tests/ensure-store.spec.ts
@@ -5,12 +5,12 @@ import {
   getSelectorMetadata,
   Selector,
   NgxsModule,
-  SelectorOptions
+  SelectorOptions,
 } from '@ngxs/store';
 import { TestBed } from '@angular/core/testing';
 
 import { SelectorMetaDataModel } from '../src/internal/internals';
-import { getRootSelectorFactory } from '../src/utils/selector-utils';
+import { getRootSelectorFactory } from '../src/selectors/selector-utils';
 import { Injectable } from '@angular/core';
 
 describe('Ensure metadata', () => {
@@ -23,7 +23,7 @@ describe('Ensure metadata', () => {
   describe('Ensure store for plugins', () => {
     @State({
       name: 'myCounter',
-      defaults: 1
+      defaults: 1,
     })
     @Injectable()
     class MyCounterState {
@@ -40,7 +40,7 @@ describe('Ensure metadata', () => {
     @State({
       name: 'count',
       defaults: 0,
-      children: [MyCounterState]
+      children: [MyCounterState],
     })
     @Injectable()
     class CountState extends MyCounterState {
@@ -58,7 +58,7 @@ describe('Ensure metadata', () => {
 
     beforeAll(() => {
       TestBed.configureTestingModule({
-        imports: [NgxsModule.forRoot([CountState, MyCounterState])]
+        imports: [NgxsModule.forRoot([CountState, MyCounterState])],
       });
     });
 
@@ -68,14 +68,14 @@ describe('Ensure metadata', () => {
         actions: {
           increment: [
             { fn: 'addOne', options: {}, type: 'increment' },
-            { fn: 'addTwo', options: {}, type: 'increment' }
+            { fn: 'addTwo', options: {}, type: 'increment' },
           ],
-          decrement: [{ fn: 'decrement', options: {}, type: 'decrement' }]
+          decrement: [{ fn: 'decrement', options: {}, type: 'decrement' }],
         },
         defaults: 0,
         path: null,
         makeRootSelector: expect.any(Function),
-        children: [MyCounterState]
+        children: [MyCounterState],
       });
     });
 
@@ -86,7 +86,7 @@ describe('Ensure metadata', () => {
         defaults: 1,
         path: null,
         makeRootSelector: expect.any(Function),
-        children: undefined
+        children: undefined,
       });
     });
 
@@ -146,7 +146,7 @@ describe('Ensure metadata', () => {
       try {
         @State({
           name: 'superCount',
-          defaults: 0
+          defaults: 0,
         })
         @Injectable()
         class SuperCountState extends MyCounterState {
@@ -172,10 +172,7 @@ describe('Ensure metadata', () => {
     });
 
     function flatString(str: string): string {
-      return str
-        .toString()
-        .replace(/\n/g, '')
-        .replace(/\s\s+/g, ' ');
+      return str.toString().replace(/\n/g, '').replace(/\s\s+/g, ' ');
     }
   });
 });
